test: split newProductForm.spec into validation + advanced files
newProductForm.spec.ts was the suite's largest single file at ~6m at
workers=1, which set the floor on whichever shard contained it
(~10m 30s slowest shard). The bin-packer can't split a single
.spec.ts file across shards, so the only way to redistribute was to
turn it into multiple files.

Following the existing announcements/ precedent
(announcements.spec.ts + announcementsNewUI.spec.ts):

  new-product-form/
  ├── newProductForm.spec.ts             KEEP — happy paths + edge cases (6 tests)
  ├── newProductFormValidation.spec.ts   NEW — negative cases (9 tests)
  ├── newProductFormAdvanced.spec.ts     NEW — product types + inventory
  │                                            + sidebar + advanced (12 tests)
  └── newProductFormPage.ts              UNCHANGED — page object

Zero test-logic changes:
 - 27 tests in, 27 tests out (verified by grep)
 - All test names preserved verbatim
 - All tags (@lite, @vendor, @pro) preserved verbatim
 - Page object untouched
 - Same beforeEach/afterEach pattern duplicated into each file
   (browser context, storage state, navigation) — IDE may flag the
   duplication; that's an accepted trade-off for shardability.

After the next CI run completes, the shard-durations-baseline
artifact will include real timings for the 3 new files; refresh
shard-durations.json from that to let the bin-packer redistribute
optimally. Expected total CI time: ~12m 31s -> ~9-10 min.
