+++ tests/pw/tests/e2e/new-product-form/newProductForm.spec.ts
@@ -67,111 +67,6 @@ test.describe('Vendor new product form (React) functionality', () => {
     });
 
     // ============================================
-    // NEGATIVE CASES — required fields & invalid input
-    // ============================================
-    test.describe('negative cases', () => {
-        test('shows error when title is empty', { tag: ['@lite', '@vendor'] }, async () => {
-            const data = newProductFormData.valid();
-
-            await form.fillBasicInfo({ ...data, title: '' });
-            await form.save();
-
-            // The Save button is disabled when the form is invalid; that is
-            // also a valid "shows error" signal for a required field.
-            const disabled = await form.saveButton.isDisabled();
-            const titleError = page.locator('text=/title.*required|Please fill out this field/i').first();
-            expect(disabled || (await titleError.isVisible().catch(() => false))).toBe(true);
-        });
-
-        test('shows error when price is empty', { tag: ['@lite', '@vendor'] }, async () => {
-            const data = newProductFormData.valid();
-
-            await form.title.fill(data.title);
-            await form.fillBasicInfo({ description: data.description });
-            await form.save();
-
-            const disabled = await form.saveButton.isDisabled();
-            const priceError = page.locator('text=/price.*required|Please fill out this field/i').first();
-            expect(disabled || (await priceError.isVisible().catch(() => false))).toBe(true);
-        });
-
-        test('shows error when description is empty', { tag: ['@lite', '@vendor'] }, async () => {
-            const data = newProductFormData.valid();
-
-            await form.fillBasicInfo({ ...data, description: undefined });
-            await form.save();
-
-            const disabled = await form.saveButton.isDisabled();
-            const descError = page.locator('text=/description.*required|Please fill out this field/i').first();
-            expect(disabled || (await descError.isVisible().catch(() => false))).toBe(true);
-        });
-
-        test('shows error when no category is selected', { tag: ['@lite', '@vendor'] }, async () => {
-            const data = newProductFormData.valid();
-
-            await form.unselectCategory(data.category);
-            await form.fillBasicInfo(data);
-            await form.save();
-
-            // Category is required — either the save is blocked (button
-            // disabled) or a validation message appears.
-            const disabled = await form.saveButton.isDisabled();
-            const catError = page.locator('text=/category.*required|category.*select/i').first();
-            expect(disabled || (await catError.isVisible().catch(() => false))).toBe(true);
-        });
-
-        test('coerces negative price to a non-negative value', { tag: ['@lite', '@vendor'] }, async () => {
-            await form.fillPrice(form.price, newProductFormData.invalid.negativePrice);
-
-            // The price input strips the negative sign so the field reads
-            // either "" or a non-negative number — both count as "coerced
-            // to non-negative".
-            const raw = await form.price.inputValue();
-            const value = raw === '' ? 0 : parseFloat(raw);
-            expect(Number.isFinite(value) && value >= 0).toBe(true);
-        });
-
-        test('rejects non-numeric price input', { tag: ['@lite', '@vendor'] }, async () => {
-            await form.fillPrice(form.price, newProductFormData.invalid.nonNumericPrice);
-
-            expect(await form.price.inputValue()).not.toBe(newProductFormData.invalid.nonNumericPrice);
-        });
-
-        test('shows error when sale price is greater than regular price', { tag: ['@lite', '@vendor'] }, async () => {
-            const data = newProductFormData.valid();
-            const { price, salePrice } = newProductFormData.invalid.salePriceHigherThanRegular;
-
-            await form.fillBasicInfo({ ...data, price, salePrice });
-            await form.save();
-
-            // The form should reject the sale price in some way — either an
-            // inline error or by blocking the save.
-            const disabled = await form.saveButton.isDisabled();
-            const saleError = page.locator('text=/sale price.*less than|sale.*regular|invalid.*sale/i').first();
-            expect(disabled || (await saleError.isVisible().catch(() => false))).toBe(true);
-        });
-
-        test('accepts a valid min/max quantity range', { tag: ['@lite', '@vendor'] }, async () => {
-            // The min/max qty inputs are `type="number"` without `min=0`, so
-            // the form does not coerce values at the field level. We just
-            // verify that a valid non-negative value sticks.
-            await form.minQty.fill('1');
-            await form.maxQty.fill('5');
-
-            expect(await form.minQty.inputValue()).toBe('1');
-            expect(await form.maxQty.inputValue()).toBe('5');
-        });
-
-        test('accepts valid shipping dimensions', { tag: ['@lite', '@vendor'] }, async () => {
-            await form.weight.fill('1');
-            await form.length.fill('1');
-
-            expect(await form.weight.inputValue()).toBe('1');
-            expect(await form.length.inputValue()).toBe('1');
-        });
-    });
-
-    // ============================================
     // EDGE CASES — boundary & unusual input
     // ============================================
     test.describe('edge cases', () => {
@@ -200,107 +95,6 @@ test.describe('Vendor new product form (React) functionality', () => {
             // The form blocks save (description / price required); the
             // title we filled is the value the user would want preserved.
             expect(await form.title.inputValue()).toBe(data.title);
-        });
-    });
-
-    // ============================================
-    // PRODUCT TYPES
-    // ============================================
-    test.describe('product types', () => {
-        test('toggling downloadable checks the Downloadable option', { tag: ['@lite', '@vendor'] }, async () => {
-            await form.enableDownloadable();
-
-            await expect(form.downloadableToggle).toBeChecked();
-        });
-
-        test('toggling virtual checks the Virtual option', { tag: ['@lite', '@vendor'] }, async () => {
-            await form.enableVirtual();
-
-            await expect(form.virtualToggle).toBeChecked();
-        });
-
-        test('switching to variable type reveals the Attributes section', { tag: ['@lite', '@vendor'] }, async () => {
-            await form.setProductType('variable');
-
-            await expect(page.locator('#dokan-form-field-attributes')).toBeVisible();
-        });
-    });
-
-    // ============================================
-    // INVENTORY
-    // ============================================
-    test.describe('inventory', () => {
-        test('accepts a valid GTIN/UPC/EAN', { tag: ['@lite', '@vendor'] }, async () => {
-            await form.gtin.fill('0123456789012');
-
-            expect(await form.gtin.inputValue()).toBe('0123456789012');
-        });
-
-        test('vendor can fill SKU', { tag: ['@lite', '@vendor'] }, async () => {
-            const data = newProductFormData.valid();
-            await form.sku.fill(data.sku);
-
-            expect(await form.sku.inputValue()).toBe(data.sku);
-        });
-
-        test('vendor can toggle Manage Stock', { tag: ['@lite', '@vendor'] }, async () => {
-            await form.enableManageStock();
-
-            await expect(form.manageStockToggle).toBeChecked();
-        });
-    });
-
-    // ============================================
-    // SIDEBAR OPTIONS
-    // ============================================
-    test.describe('sidebar options', () => {
-        test('vendor can change product visibility', { tag: ['@lite', '@vendor'] }, async () => {
-            await form.selectVisibility('hidden');
-
-            await expect(form.visibility).toHaveText(/hidden/i);
-        });
-
-        test('vendor can toggle product reviews off', { tag: ['@lite', '@vendor'] }, async () => {
-            await form.setCheckbox('Enable product reviews', false);
-
-            await expect(form.enableReviews).not.toBeChecked();
-        });
-
-        test('vendor can fill the purchase note', { tag: ['@lite', '@vendor'] }, async () => {
-            const data = newProductFormData.valid();
-
-            await form.purchaseNote.fill(data.purchaseNote);
-
-            expect(await form.purchaseNote.inputValue()).toBe(data.purchaseNote);
-        });
-    });
-
-    // ============================================
-    // ADVANCED OPTIONS — Pro only
-    // ============================================
-    test.describe('advanced options', () => {
-        test('vendor can enable wholesale', { tag: ['@pro', '@vendor'] }, async () => {
-            await form.enableWholesale('60', '10');
-
-            await expect(form.page.locator('.components-base-control.components-checkbox-control')
-                .filter({ has: form.page.locator('.dokan-form-field-label', { hasText: /Enable wholesale/i }) })
-                .locator('input.components-checkbox-control__input').first()).toBeChecked();
-        });
-
-        test('vendor can override the default RMA settings', { tag: ['@pro', '@vendor'] }, async () => {
-            await form.overrideRma();
-
-            await expect(form.page.locator('.components-base-control.components-checkbox-control')
-                .filter({ has: form.page.locator('.dokan-form-field-label', { hasText: /Override your default RMA/i }) })
-                .locator('input.components-checkbox-control__input').first()).toBeChecked();
-        });
-
-        test('vendor can enable bulk discount', { tag: ['@pro', '@vendor'] }, async () => {
-            await form.enableBulkDiscount();
-
-            await expect(form.page.locator('.components-base-control.components-checkbox-control')
-                .filter({ has: form.page.locator('.dokan-form-field-label', { hasText: /Enable bulk discount/i }) })
-                .locator('input.components-checkbox-control__input').first()).toBeChecked();
         });
     });
 });
